Move combine to separate module
